[Refactor]: Rearrang folder structure for reporter
