+++ src/app/app-routing.module.ts
@@ -4,8 +4,7 @@ import { Routes, RouterModule } from '@angular/router';
 const routes: Routes = [
   {
     path: 'report-dashboard',
-    loadChildren: () =>
-      import('./reporters-dashboard/reporters-dashboard.module').then((m) => m.ReportersDashboardModule)
+    loadChildren: () => import('./reporter/reporters.module').then((m) => m.ReportersDashboardModule)
   },
   {
     path: 'auth',
