+++ src/app/reporter/reporters-dashboard/reporters-dashboard.component.scss
