+++ src/app/reporter/reporters-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { ReportersDashboardComponent } from './reporters-dashboard.component';
+import { ReportersDashboardComponent } from './reporters-dashboard/reporters-dashboard.component';
 
 const routes: Routes = [{ path: '', component: ReportersDashboardComponent }];
