+++ src/app/reporter/reporters.module.ts
@@ -2,8 +2,8 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 
-import { ReportersDashboardRoutingModule } from './reporters-dashboard-routing.module';
-import { ReportersDashboardComponent } from './reporters-dashboard.component';
+import { ReportersDashboardRoutingModule } from './reporters-routing.module';
+import { ReportersDashboardComponent } from './reporters-dashboard/reporters-dashboard.component';
 import { SharedModule } from '../shared/shared.module';
 
 @NgModule({
